add map to get display name in leaderboard
