+++ frontend/src/App.tsx
@@ -10,7 +10,7 @@ import AuthCallback from "./components/AuthCallback";
 import ProtectedRoute from "./components/ProtectedRoute";
 import Chat from "./components/chat/Chat";
 import GameContainer from "./components/game/GameContainer";
-import Leaderboard from "./components/game/Leaderboard";
+import Leaderboard from "./components/leaderboard/Leaderboard";
 
 const App: React.FC = () => {
   return (
