+++ package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "@types/node": "^22.13.10"
+  }
+}
